requirements.txt
aioredis==1.3.1
amqp==2.5.2
asgi-redis==1.4.3
asgiref==3.2.7
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.4.2
Automat==20.2.0
autopep8==1.5.1
billiard==3.6.3.0
celery==4.4.2
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
colorama==0.4.3
constantly==15.1.0
cryptography==2.9
daphne==2.5.0
Django==2.2
django-cors-headers==3.2.1
django-filter==2.2.0
django-webpack-loader==0.7.0
djangorestframework==3.11.0
hiredis==1.0.1
hyperlink==19.0.0
idna==2.9
importlib-metadata==1.6.0
incremental==17.5.0
isort==4.3.21
jsonfield==3.1.0
kombu==4.6.8
lazy-object-proxy==1.4.3
mccabe==0.6.1
msgpack==0.6.2
msgpack-python==0.5.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.20
PyHamcrest==2.0.2
pylint==2.4.4
pyOpenSSL==19.1.0
pytz==2019.3
redis==3.5.2
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
typed-ast==1.4.1
vine==1.3.0
wrapt==1.11.2
yapf==0.29.0
zipp==3.1.0
zope.interface==5.1.0
